rename repo to service folder
