+++ coffee/service/db.go
@@ -2,7 +2,7 @@
 // versions:
 //   sqlc v1.16.0
 
-package repository
+package service
 
 import (
 	"context"
