+++ coffee/service/models.go
@@ -2,7 +2,7 @@
 // versions:
 //   sqlc v1.16.0
 
-package repository
+package service
 
 import (
 	"database/sql"
